+++ README.md
@@ -58,8 +58,8 @@
 - Отфильтрованный Список Запросов пагинируется по принципу, указанному в секции "Список Запросов".
 
 ### Карточка запроса
-- Картинка карточки зависит от параметров `requesterType` и `helpType`, подробности в дизайне, в категории `UI Elements`.
 
+- Картинка карточки зависит от параметров `requesterType` и `helpType`, подробности в дизайне, в категории `UI Elements`.
 
 ## Страница Запроса
 
@@ -80,7 +80,7 @@
 
 ## Страница Профиля
 
--  Пользователь по умолчанию при открытии страницы видит вкладку "Личные данные".
+- Пользователь по умолчанию при открытии страницы видит вкладку "Личные данные".
 - Данные для вкладок "Личные данные" и "Контакты" загружаются API вызовом `LoadUserInfo`.
 - По нажатию на кнопку "Выйти из аккаунта" происходит действие аналогичное нажатию на пункт "Выйти" Меню профиля в шапке (смотри секцию Меню профиля в шапке).
