+++ src/auth/AuthProvider.tsx
@@ -1,6 +1,5 @@
-import { useState, useContext } from "react";
+import { useState, useContext, createContext } from "react";
 import { useNavigate } from "react-router-dom";
-import { createContext } from "react";
 
 interface AuthContextType {
   loggedIn: boolean;
