+++ src/components/login-forms/testProfiles/testProfilesData.ts
@@ -1,0 +1,23 @@
+interface TestsProfilesData {
+  name: string;
+  login: string;
+  password: string;
+}
+
+export const testsProfilesData: TestsProfilesData[] = [
+  {
+    name: "Первый пользователь",
+    login: "testUser18@test.com",
+    password: "password18",
+  },
+  {
+    name: "Второй пользователь",
+    login: "testUser19@test.com",
+    password: "password19",
+  },
+  {
+    name: "Третий пользователь",
+    login: "testUser20@test.com",
+    password: "password20",
+  },
+];
